+++ tests/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url, include
+
+
+urlpatterns = [
+    url(r'^statusboard/', include('statusboard.urls')),
+]
